Added Destination Class
I added  a class for destinations and I also updated the tour class
with more methods and ways putting all of the info into a string to be
used in the database. It can also parse the string of information. This
is preliminary so it may not be perfect yet. We can discuss all of the
specifications next time we meet.