Finished method to put markers on
I added a method that will take a tour and then get all of the
destinations associated with it and put them on the map and then move
the camera to the given primary destination associated with the tour.
The set current location method also uses a different color now.